requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
motor==3.7.1
jinja2==3.1.6
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
reportlab==4.4.3
google-generativeai==0.8.5
email-validator==2.2.0
pydantic-settings==2.10.1
